+++ src/components/Inputs/SelectInputLabel/index.ts
@@ -1,0 +1,4 @@
+import SelectInputLabel from "./SelectInputLabel";
+
+export default SelectInputLabel;
+export * from "./SelectInputLabel";